+++ .idea/compiler.xml
@@ -4,6 +4,7 @@
     <bytecodeTargetLevel target="1.7">
       <module name="PocketChef.app" target="11" />
       <module name="PocketChef.data" target="11" />
+      <module name="PocketChef.presentation" target="11" />
     </bytecodeTargetLevel>
   </component>
 </project>
