+++ .idea/misc.xml
@@ -1,5 +1,21 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
+  <component name="DesignSurface">
+    <option name="filePathToZoomLevelMap">
+      <map>
+        <entry key="..\:/Users/EXTREMEX/AndroidStudioProjects/PocketChef/app/src/main/res/layout/activity_main.xml" value="0.3546875" />
+        <entry key="..\:/Users/EXTREMEX/AndroidStudioProjects/PocketChef/app/src/main/res/layout/fragment_dashboard.xml" value="0.3546875" />
+        <entry key="..\:/Users/EXTREMEX/AndroidStudioProjects/PocketChef/app/src/main/res/layout/fragment_home.xml" value="0.17604166666666668" />
+        <entry key="..\:/Users/EXTREMEX/AndroidStudioProjects/PocketChef/app/src/main/res/layout/fragment_notifications.xml" value="0.17604166666666668" />
+        <entry key="..\:/Users/EXTREMEX/AndroidStudioProjects/PocketChef/presentation/src/main/res/drawable/ic_home_black_24dp.xml" value="0.1845" />
+        <entry key="..\:/Users/EXTREMEX/AndroidStudioProjects/PocketChef/presentation/src/main/res/layout/activity_main.xml" value="0.3546875" />
+        <entry key="..\:/Users/EXTREMEX/AndroidStudioProjects/PocketChef/presentation/src/main/res/layout/fragment_dashboard.xml" value="0.17604166666666668" />
+        <entry key="..\:/Users/EXTREMEX/AndroidStudioProjects/PocketChef/presentation/src/main/res/layout/fragment_home.xml" value="0.17604166666666668" />
+        <entry key="..\:/Users/EXTREMEX/AndroidStudioProjects/PocketChef/presentation/src/main/res/layout/fragment_notifications.xml" value="0.3546875" />
+        <entry key="..\:/Users/EXTREMEX/AndroidStudioProjects/PocketChef/presentation/src/main/res/menu/bottom_nav_menu.xml" value="0.17604166666666668" />
+      </map>
+    </option>
+  </component>
   <component name="ProjectRootManager" version="2" languageLevel="JDK_1_7" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
