+++ presentation/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id 'com.android.application'
+    id 'com.android.library'
     id 'org.jetbrains.kotlin.android'
 }
 
@@ -7,13 +7,13 @@ android {
     compileSdk 32
 
     defaultConfig {
-        applicationId "com.example.pocketchef"
         minSdk 29
         targetSdk 32
         versionCode 1
         versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles "consumer-rules.pro"
     }
 
     buildTypes {
@@ -36,9 +36,8 @@ android {
 
 dependencies {
 
-    implementation project(path: ':data')
     implementation project(path: ':domain')
-    implementation project(path: ':presentation')
+    implementation project(path: ':data')
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.4.1'
     implementation 'com.google.android.material:material:1.5.0'
