+++ presentation/src/main/java/com/example/pocketchef/presentation/ui/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.pocketchef
+package com.example.pocketchef.presentation.ui
 
 import android.os.Bundle
 import com.google.android.material.bottomnavigation.BottomNavigationView
@@ -7,7 +7,8 @@ import androidx.navigation.findNavController
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
-import com.example.pocketchef.databinding.ActivityMainBinding
+import com.example.pocketchef.presentation.R
+import com.example.pocketchef.presentation.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
