+++ presentation/src/main/java/com/example/pocketchef/presentation/ui/dashboard/DashboardViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.pocketchef.ui.notifications
+package com.example.pocketchef.presentation.ui.dashboard
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class DashboardViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is notifications Fragment"
+        value = "This is dashboard Fragment"
     }
     val text: LiveData<String> = _text
 }
