+++ presentation/src/main/java/com/example/pocketchef/presentation/ui/home/HomeViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.pocketchef.ui.dashboard
+package com.example.pocketchef.presentation.ui.home
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class HomeViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is dashboard Fragment"
+        value = "This is home Fragment"
     }
     val text: LiveData<String> = _text
 }
