+++ presentation/src/main/java/com/example/pocketchef/presentation/ui/notifications/NotificationsViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.pocketchef.ui.home
+package com.example.pocketchef.presentation.ui.notifications
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class HomeViewModel : ViewModel() {
+class NotificationsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
+        value = "This is notifications Fragment"
     }
     val text: LiveData<String> = _text
 }
