+++ presentation/src/main/res/drawable/ic_dashboard_black_24dp.xml
