+++ presentation/src/main/res/layout/fragment_notifications.xml
@@ -4,10 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.dashboard.DashboardFragment">
+    tools:context="com.example.pocketchef.presentation.ui.notifications.NotificationsFragment">
 
     <TextView
-        android:id="@+id/text_dashboard"
+        android:id="@+id/text_notifications"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginEnd="8dp"
