+++ settings.gradle
@@ -16,3 +16,4 @@ rootProject.name = "PocketChef"
 include ':app'
 include ':data'
 include ':domain'
+include ':presentation'
